fix(playerbot): Move ObjectiveState and AuctionSession to interface, fix BuildOptions
- Moved ObjectiveState struct to IObjectiveTracker.h interface with full definition
- Moved AuctionSession struct to IAuctionHouse.h interface
- Fixed SpellPacketBuilder BuildOptions default argument (overload instead)
- Fixed HunterAI malformed include
- Used static_cast for AuctionActionType default values

Remaining compilation errors:
- ObjectiveTracker covariant return type still persisting (needs investigation)
- SpellPacketBuilder line 228 still has default member initializer issue
- PaladinAI/HunterAI missing return statements
- AuctionMetrics needs same treatment as other structs
- PriestAI malformed include
