+++ src/modules/Playerbot/AI/ClassAI/Hunters/HunterAI.cpp
@@ -30,7 +30,7 @@
 #include <chrono>
 #include "../../../Spatial/SpatialGridManager.h"
 #include "../../../Spatial/SpatialGridQueryHelpers.h"  // PHASE 5F: Thread-safe queries
-#include "../../../Movement/UnifiedMovementCoordinator.h  // Phase 2: Unified movement system"
+#include "../../../Movement/UnifiedMovementCoordinator.h"
 #include "../../../Movement/Arbiter/MovementPriorityMapper.h"
 #include "../../BotAI.h"
 #include "UnitAI.h"
